requirements.txt
# Core App
streamlit==1.35.0
python-dotenv==1.0.1
requests==2.31.0
tqdm==4.66.4

# Neo4j
neo4j==5.14.0

# LangChain
openai>=1.86.0,<2.0.0
langchain-openai>=0.3.27
llama-index>=0.10.0


# PDF Processing
PyMuPDF==1.24.2

# Vector Store (Chroma Optional, not essential for Graph-only RAG)
chromadb[duckdb]==0.5.0

# MongoDB for Chat Memory
pymongo==4.6.3

# RAGAS Evaluation
ragas[all]==0.0.16
datasets==2.19.1
evaluate==0.4.2
pillow
pytesseract